update usages of certificate and key endpoints
